Show progress prettier.

+++ src/integration/java/com/infodesire/jglu/busdata/StoreLargeNumberOfBusObjectsTest.java
@@ -75,8 +75,8 @@ public class StoreLargeNumberOfBusObjectsTest {
 
             if( i % 1000 == 0 ) {
                 long t1 = System.currentTimeMillis();
-                double rate = Math.round( ( (double) c0 / (double) ( t1 - t0 ) ) * 1000 );
-                double percent = Math.round( ( (double) i / (double) OBJECT_COUNT ) * 100.0 );
+                int rate = (int) Math.round( ( (double) c0 / (double) ( t1 - t0 ) ) * 1000 );
+                int percent = (int) Math.round( ( (double) i / (double) OBJECT_COUNT ) * 100.0 );
                 System.out.println( "#" + i + "/" + OBJECT_COUNT + " (" + percent + " %) Create rate: " + rate + " objects/s" );
                 t0 = System.currentTimeMillis();
                 c0 = 0;
