+++ src/main/java/com/infodesire/jglu/Main.java
@@ -165,8 +165,8 @@ public class Main {
                 long t1 = System.currentTimeMillis();
                 long time = t1 - t0;
                 long count = index - c0;
-                double rate = Math.round( ( (double) count / (double) time ) * 1000 );
-                double percent = Math.round( ( (double) index / (double) total ) * 100.0 );
+                int rate = (int) Math.round( ( (double) count / (double) time ) * 1000 );
+                int percent = (int) Math.round( ( (double) index / (double) total ) * 100.0 );
                 System.out.println( index + "/" + total + " (" + percent + " %) " + rate + " del/s" );
                 t0 = t1;
                 c0 = index;
